requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
streamlit>=1.52.0
requests>=2.32.0
python-dotenv>=1.0.0
